be2/requirements.txt
mysql-connector-python
python-dotenv
mysql-connector-python
